% Generated by roxygen2 (4.0.0): do not edit by hand
\name{color.choose}
\alias{color.choose}
\title{Choose colors interactively from a grid}
\usage{
color.choose(detail = 100, pick = TRUE)
}
\arguments{
\item{detail}{an integer specifying the smoothness of color chart.
In essence it gives the NxN dimensions of the displayed color grid,
which in turn dictates how many colors are displayed. Large values may slow
down the process. Defaults to 200 (40 000 colors).}

\item{pick}{logical, whether or not the user should be allowed to select a
color. If pick == FALSE locations of the colors are returned.  This may be
useful if you want to pick colors from within another
function.}
}
\description{
color.choose
}
\details{
color.choose helps to quickly select colors from a grid for usage in R
graphics. See the example below for usage details. This version gives a
much prettier grid of colors in a range of hue and lightness values compared
to the simpler color.pick function.
}
\examples{
# pick a color
mycol<-color.choose()
plot(rnorm(100),col=mycol)
}
\author{
Marco D. Visser
}

