% Generated by roxygen2 (4.0.0): do not edit by hand
\name{color.pick}
\alias{color.pick}
\title{Pick colors interactively from a named list}
\usage{
color.pick()
}
\description{
color.pick
}
\details{
color.pick allows a user to quickly select a color from
a list of R color names. This function is depreciated and
may disappear in subsequent version of the package.
}
\examples{
# pick a color
mycol<-color.pick()
plot(rnorm(100),col=mycol)
}
\author{
Marco D. Visser
}

