% Generated by roxygen2 (4.0.0): do not edit by hand
\name{palette.picker}
\alias{palette.picker}
\title{Choose a palette interactively from a grid}
\usage{
palette.picker(n = 6, detail = 100)
}
\arguments{
\item{n}{an integer specifying how many colors to make a pallete from.}

\item{detail}{an integer specifying the smoothness of color chart.
In essence it gives the NxN dimensions of the displayed color grid,
which in turn dictates how many colors are displayed. Large values may slow
down the process. Defaults to 100 (10 000 colors).}
}
\description{
palette.picker
}
\details{
palette.picker aids in the selection of colors from a grid
and builds a palette with colorRampPalette for usage in R
graphics.  See the example below for usage details.
}
\examples{
load volcano data
data(volcano)
#Select 5 colors to make a palette
Mypalette<-palette.picker(n=5)
create an image with the new palette
image(volcano,col=MyPallete(1000))
}
\author{
Marco D. Visser
}

