% Generated by roxygen2 (4.0.0): do not edit by hand
\name{color.wheel}
\alias{color.wheel}
\title{Choose a color interactively from a color wheel}
\usage{
color.wheel(detail = 200)
}
\arguments{
\item{detail}{an integer specifying the smoothness of color chart.
In essence this specifies the number of colors plotted on the wheel
(large values may slow the process down considerably).}
}
\description{
color.wheel
}
\details{
color.wheel allows greater detail in color selection compared to
 color.choose(). The user first chooses the color and then the
 lightness of the color.
}
\examples{
# Get more detail in choice with the color wheel
mycol<-color.wheel()
plot(rnorm(100),col=mycol)
}
\author{
Marco D. Visser
}

